+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add duration properties to the `Pledge` class
 * Add `Blast.REPLACE_OPEN_TAG_NEWLINE` symbol for use in `String#stripTags()`
 * Add `Deck#remove(key)` method
+* Fix `String#romanize()` stripping diacritics before being able to replace them
 
 ## 0.6.3 (2018-12-06)